requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Weather API
requests==2.31.0

# AI/LLM
openai==1.3.5
langchain==0.0.340
langchain-openai==0.0.2

# Data validation
pydantic==2.5.0

# Utilities
python-multipart==0.0.6